feat(componentbase): new component base for component that need the select button such as ghget and ghtidyup
